Add files via upload
The box contains multi-colored beads (or beads). All beads have the same shape, size and weight. Beads can be one of N different colors. There are a lot of beads in each box in each color.
It is required to determine the minimum number of beads that can be pulled out of the box without looking so that between them are guaranteed to be two beads of the same color.